requirements.txt
-r requirements/staging.txt
arrow==0.10.0
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.5
django-disqus==0.5
django-emoticons==1.2
django-forms-bootstrap==3.1.0
django-paypal==0.4.1
django-tinymce==2.6.0
funcsigs==1.0.2
gunicorn==19.7.1
idna==2.6
mock==2.0.0
olefile==0.44
pbr==3.1.1
Pillow==4.3.0
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
six==1.11.0
urllib3==1.22
dj-database-url==0.2.1